Extract shared startup tilt observer and detrender defaults
